Update pricing accounts
