+++ anchor/programs/glam/src/state/mod.rs
@@ -12,3 +12,5 @@ pub use assets::*;
 
 pub mod acl;
 pub use acl::*;
+
+pub mod pyth_price;
